+++ src/api/execute-all.js
@@ -1,0 +1,9 @@
+
+import getRefs from '../core/ref/get-refs';
+
+export default ()=>{
+  const refs = getRefs();
+  refs.forEach(ref=>{
+    if(ref.$$execute)ref.$$execute();
+  });
+}